feat: migrate to pnpm with scripts
