--- .eslintrc.js
@@ -1,67 +1,0 @@
-module.exports = {
-	env: {
-		browser: true,
-		commonjs: true,
-		es2020: true,
-		node: true,
-	},
-	extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
-	parser: '@typescript-eslint/parser',
-	parserOptions: {
-		ecmaVersion: 12,
-	},
-	plugins: ['@typescript-eslint'],
-	ignorePatterns: '*.test.js',
-	rules: {
-		'@typescript-eslint/no-var-requires': 0,
-		'brace-style': ['error', '1tbs', { allowSingleLine: true }],
-		'comma-dangle': ['error', 'always-multiline'],
-		'comma-spacing': 'error',
-		'comma-style': 'error',
-		curly: ['error', 'multi-line', 'consistent'],
-		'dot-location': ['error', 'property'],
-		'handle-callback-err': 'off',
-		indent: ['error', 'tab'],
-		'max-nested-callbacks': ['error', { max: 4 }],
-		'max-statements-per-line': ['error', { max: 2 }],
-		'no-console': 'error',
-		'no-empty-function': 'error',
-		'no-floating-decimal': 'error',
-		'no-inline-comments': 'error',
-		'no-multi-spaces': 'error',
-		'no-multiple-empty-lines': ['error', { max: 2, maxEOF: 1, maxBOF: 0 }],
-		'no-shadow': ['error', { allow: ['err', 'resolve', 'reject'] }],
-		'no-trailing-spaces': ['error', { 'skipBlankLines': true, 'ignoreComments': true }],
-		'no-var': 'error',
-		'object-curly-spacing': ['error', 'always'],
-		'prefer-const': 'error',
-		quotes: ['error', 'single'],
-		semi: ['error', 'always'],
-		'space-before-blocks': 'error',
-		'space-before-function-paren': [
-			'error',
-			{
-				anonymous: 'never',
-				named: 'never',
-				asyncArrow: 'always',
-			},
-		],
-		'space-in-parens': 'error',
-		'space-infix-ops': 'error',
-		'space-unary-ops': 'error',
-		'spaced-comment': 'error',
-		yoda: 'error',
-		'@typescript-eslint/explicit-function-return-type': 0,
-		'@typescript-eslint/explicit-member-accessibility': 0,
-		'@typescript-eslint/indent': 0,
-		'@typescript-eslint/member-delimiter-style': 0,
-		'@typescript-eslint/no-explicit-any': 0,
-		'@typescript-eslint/no-use-before-define': 0,
-		'@typescript-eslint/no-unused-vars': [
-			2,
-			{
-				'argsIgnorePattern': '^_',
-			},
-		],
-	},
-};
