+++ .github/CODEOWNERS
@@ -1,4 +1,4 @@
 # This is a comment.
 # Each line is a file pattern followed by one or more owners.
-* @nonsensecodes @amaredeus @banklessdao/devops
+@amaredeus @banklessdao/devops
 /.github/ @banklessdao/devops
