+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 1. Repo initialized
 2. Upgrade mongodb -> 4.x, add sentry, remove dd, update deps
 3. Update github action with sentry
+4. Migrate yarn -> pnpm
