+++ package.json
@@ -7,13 +7,15 @@
     "scripts": {
         "build": "tsc -p tsconfig.json",
         "test": "jest",
-        "qa": "node -r dotenv/config --trace-warnings dist/app/app.js dotenv_config_path=.env.qa",
-        "prod": "node -r dotenv/config --trace-warnings dist/app/app.js dotenv_config_path=.env.prod",
-        "prestart": "yarn install && yarn build",
-        "pretest": "yarn install && yarn build",
-        "start": "node --trace-warnings -r dotenv/config dist/app/app.js",
-        "lint": "eslint ./src/app/**/*.ts --ext .ts",
-        "format": "eslint ./src/app/**/*.ts --ext .ts --fix",
+        "start:qa": "node -r dotenv/config --trace-warnings dist/app/App.js dotenv_config_path=.env.qa",
+        "start:prod": "node -r dotenv/config --trace-warnings dist/app/App.js dotenv_config_path=.env.prod",
+        "prestart": "pnpm install && pnpm build && pnpm lint",
+        "pretest": "pnpm install && pnpm build",
+        "dev": "pnpm prestart && pnpm start",
+        "start": "node --trace-warnings -r dotenv/config dist/app/App.js",
+        "force-start": "tsc -p tsconfig.json && node -r dotenv/config dist/app/App.js",
+        "lint": "eslint . --ext .ts",
+        "format": "eslint . --ext .ts --fix",
         "watch": "tsc -p tsconfig.json -w"
     },
     "repository": {
@@ -33,32 +35,33 @@
     },
     "homepage": "https://github.com/BanklessDAO/discord-bot/blob/main/README.md",
     "dependencies": {
-        "@logdna/logger": "^2.6.0",
-        "@sentry/integrations": "^6.16.1",
-        "@sentry/node": "^6.16.1",
-        "@sentry/tracing": "^6.16.1",
-        "@types/node": "^16.7.1",
-        "dayjs": "^1.10.7",
-        "dd-trace": "^1.5.1",
+        "@logdna/logger": "^2.6.1",
+        "@sentry/integrations": "^6.19.6",
+        "@sentry/node": "^6.19.6",
+        "@sentry/tracing": "^6.19.6",
+        "@types/node": "^16.11.26",
+        "dayjs": "^1.11.0",
+        "dd-trace": "^1.7.1",
         "discord.js": "^13.6.0",
-        "dotenv": "^14.2.0",
+        "dotenv": "^14.3.2",
         "form-data": "^4.0.0",
-        "mongodb": "^4.3.0",
+        "mongodb": "^4.5.0",
         "p-queue": "^6.6.2",
-        "slash-create": "^5.0.3",
+        "slash-create": "^5.4.1",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
-        "@shelf/jest-mongodb": "^2.2.0",
-        "@types/jest": "^27.4.0",
-        "@typescript-eslint/eslint-plugin": "^5.10.0",
-        "@typescript-eslint/parser": "^5.10.0",
+        "@shelf/jest-mongodb": "^2.2.1",
+        "@types/jest": "^27.4.1",
+        "@typescript-eslint/eslint-plugin": "^5.18.0",
+        "@typescript-eslint/parser": "^5.18.0",
         "builder-pattern": "^1.3.0",
-        "eslint": "^8.7.0",
-        "jest": "^27.4.7",
-        "ts-jest": "^27.1.3",
-        "ts-loader": "^9.2.6",
-        "typescript": "^4.5.4"
+        "eslint": "^8.12.0",
+        "jest": "^27.5.1",
+        "ts-jest": "^27.1.4",
+        "ts-loader": "^9.2.8",
+        "typescript": "^4.6.3",
+        "webpack": "^5.72.0"
     },
     "engines": {
         "yarn": "1.x",
