+++ src/app/errors/EarlyTermination.ts
@@ -1,9 +1,9 @@
 export default class EarlyTermination extends Error {
 
-	constructor(message: string) {
-		super(message);
+  constructor(message: string) {
+    super(message);
 
-		Object.setPrototypeOf(this, EarlyTermination.prototype);
-	}
+    Object.setPrototypeOf(this, EarlyTermination.prototype);
+  }
 
 }
