+++ src/app/service/constants/apiKeys.ts
@@ -1,11 +1,11 @@
 const apiKeys = Object.freeze({
-	DISCORD_BOT_ID: process.env.DISCORD_BOT_APPLICATION_ID,
+  DISCORD_BOT_ID: process.env.DISCORD_BOT_APPLICATION_ID,
 	
-	logDNAToken: process.env.LOGDNA_TOKEN as string,
-	logDNAAppName: process.env.LOGDNA_APP_NAME,
-	logDNADefault: process.env.LOGDNA_DEFAULT_LEVEL,
+  logDNAToken: process.env.LOGDNA_TOKEN as string,
+  logDNAAppName: process.env.LOGDNA_APP_NAME,
+  logDNADefault: process.env.LOGDNA_DEFAULT_LEVEL,
 	
-	sentryDSN: process.env.SENTRY_IO_DSN,
+  sentryDSN: process.env.SENTRY_IO_DSN,
 });
 
 export default apiKeys;
