+++ src/app/service/constants/channelIds.ts
@@ -1,4 +1,4 @@
 // TODO: add channel ID
 export default Object.freeze({
-	channelExample: 'channel-id-here',
+  channelExample: 'channel-id-here',
 });
