+++ src/app/service/constants/discordServerIds.ts
@@ -1,4 +1,4 @@
 // TODO: fill out discordServerId
 export default Object.freeze({
-	banklessDAO: 'discord-server-id-here',
+  banklessDAO: 'discord-server-id-here',
 });
